+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.0.0] (Unreleased)
 
+- fix: make `Ductwork::Job#execution_crashed!` idempotent to prevent race condition with the reaper
 - fix: prevent possible double job execution by only wrapping user code in rescue
 - fix: atomically release branches conditioned on claim token
 - fix: fail abandoned availabilities and executions when reaping jobs